Conflicts:
	src/main/webapp/js/ezApprovalG/conn_HWP.js

+++ src/main/java/egovframework/ezEKP/ezCircular/web/EzCircularController.java
@@ -1,20 +1,34 @@
 package egovframework.ezEKP.ezCircular.web;
 
+import java.io.BufferedInputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.net.URLEncoder;
+import java.nio.charset.Charset;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.Properties;
 import java.util.UUID;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
 
 import javax.annotation.Resource;
 import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeUtility;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.lang3.StringUtils;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -285,12 +299,14 @@ public class EzCircularController extends EgovFileMngUtil {
 		String realPath = commonUtil.getRealPath(request);
 		String uploadFilePath = commonUtil.getUploadPath("upload_circular.ROOT", userInfo.getTenantId());
 		
-		if (request.getParameter("filePath") != null && !request.getParameter("filePath").equals("")) {
-			filePath = request.getParameter("filePath"); 
-        }
-		
 		if (request.getParameter("fileName") != null && !request.getParameter("fileName").equals("")) {
 			fileName = request.getParameter("fileName"); 
+		}
+
+		if (request.getParameter("filePath") != null && !request.getParameter("filePath").equals("")) {
+			filePath = request.getParameter("filePath");
+			//2018-07-09 김보미 - 파일부분 수정
+			filePath = filePath + fileName;
         }
 
 		if (request.getParameter("circularFileID") != null && !request.getParameter("circularFileID").equals("")) {
@@ -1109,8 +1125,11 @@ public class EzCircularController extends EgovFileMngUtil {
 			StringBuilder strAttach = new StringBuilder();
 			strAttach.append("<ROOT><NODES>");
 			
-			for (CircularAttachVO attach : attachList) {				
-				strAttach.append("<DATA><![CDATA[" + attach.getFilePath().split("/")[2] + "]]></DATA>");
+			for (CircularAttachVO attach : attachList) {
+				//2018-07-06 김보미 - 파일부분수정(data의 값 수정)
+//				strAttach.append("<DATA><![CDATA[" + attach.getFilePath().split("/")[2] + "]]></DATA>");
+				String data = attach.getFilePath().split("/")[2].replace(attach.getFileName(),"");
+				strAttach.append("<DATA><![CDATA[" + data + "]]></DATA>");
 				strAttach.append("<DATA2><![CDATA[" + attach.getFileName() + "]]></DATA2>");
 				strAttach.append("<DATA3><![CDATA[" + attach.getFileSize() + "]]></DATA3>");
 				strAttach.append("<DATA4><![CDATA[]]></DATA4>");
@@ -1364,6 +1383,11 @@ public class EzCircularController extends EgovFileMngUtil {
         		
         		String fileSize = commonUtil.byteCalculation(Long.toString(avo.getFileSize()));
         		avo.setFileTranSize(fileSize);
+        		
+        		//2018-07-09 김보미 - filePath수정
+        		String filePath = avo.getFilePath();
+        		filePath = filePath.substring(0, filePath.indexOf(avo.getFileName()));
+        		avo.setFilePath(filePath);
         	}
         	
         	model.addAttribute("attachList", aList);
@@ -1546,20 +1570,22 @@ public class EzCircularController extends EgovFileMngUtil {
             String extend = pFileName[i].substring(pFileName[i].lastIndexOf(".") + 1);
             String newFileName = pUploadSN[i];
             
-            if (useExtension.toLowerCase().indexOf(extend.toLowerCase()) == -1 && !useExtension.equals("*")) {           	
-				strXML.append("<DATA><![CDATA[" + newFileName + ";" + pFileName[i] + "]]></DATA>");
-				strXML.append("<DATA2><![CDATA[" + pFileName[i] + "]]></DATA2>");
-				strXML.append("<DATA3><![CDATA[" + fileSize[i] + "]]></DATA3>");
+            if (useExtension.toLowerCase().indexOf(extend.toLowerCase()) == -1 && !useExtension.equals("*")) {
+            	//2018-07-06 김보미
+				//strXML.append("<DATA><![CDATA[" + newFileName + ";" + pFileName[i] + "]]></DATA>");
+				strXML.append("<DATA><![CDATA[" + newFileName + "]]></DATA>");//UUID
+				strXML.append("<DATA2><![CDATA[" + pFileName[i] + "]]></DATA2>");//파일명
+				strXML.append("<DATA3><![CDATA[" + fileSize[i] + "]]></DATA3>");//파일사이즈
 				strXML.append("<DATA4><![CDATA[]]></DATA4>");
 				strXML.append("<DATA5><![CDATA[denied]]></DATA5>");
             } else {
-//            	if (mode.equals("temp")) {
-            	writeUploadedFile(multiFile.get(i), newFileName + ";" + pFileName[i], pDirPath + "tempUploadFile");            		
-//            	} else {
-//            		writeUploadedFile(multiFile.get(i), newFileName + ";" + pFileName[i], pDirPath + "uploadFile" + commonUtil.separator + circularID + "_uploadFile");
-//            	}
+            	//2018-07-06 김보미 - 불필요한 주석문 제거 및 파일명 변경
+            	//writeUploadedFile(multiFile.get(i), newFileName + ";" + pFileName[i], pDirPath + "tempUploadFile");            		
+            	writeUploadedFile(multiFile.get(i), newFileName + pFileName[i], pDirPath + "tempUploadFile");            		
             	
-				strXML.append("<DATA><![CDATA[" + newFileName + ";" + pFileName[i] + "]]></DATA>");
+            	//2018-07-06 김보미
+				//strXML.append("<DATA><![CDATA[" + newFileName + ";" + pFileName[i] + "]]></DATA>");
+				strXML.append("<DATA><![CDATA[" + newFileName + "]]></DATA>");
 				strXML.append("<DATA2><![CDATA[" + pFileName[i] + "]]></DATA2>");
 				strXML.append("<DATA3><![CDATA[" + fileSize[i] + "]]></DATA3>");
 				strXML.append("<DATA4><![CDATA[]]></DATA4>");
@@ -1609,17 +1635,33 @@ public class EzCircularController extends EgovFileMngUtil {
 		
 		logger.debug("filePath : " + filePath);
 
-		if (fileList.length() != 0) {
-			String[] data = fileList.split(","); 
-			
-			for (int i=0; i<data.length; i++) {
-				String sGUID = data[i].split(";")[0];
-				String fileName = data[i].split(";")[1];
+		//2018-07-06 김보미 - 파일부분 수정
+//		if (fileList.length() != 0) {
+//			String[] data = fileList.split(","); 
+//			
+//			for (int i=0; i<data.length; i++) {
+//				String sGUID = data[i].split(";")[0];
+//				String fileName = data[i].split(";")[1];
+//				
+//				File file = new File(pDirPath + commonUtil.separator + filePath + commonUtil.separator + sGUID + ";" + fileName);
+//				
+//				file.delete();
+//			}			
+//		}
+		JSONParser jp = new JSONParser();
+		JSONArray jsonArr = (JSONArray)jp.parse(fileList);
+		
+		if (fileList.length() != 0 && !fileList.equals("")) {
+			for (int j = 0; j < jsonArr.size(); j++) {
+				JSONObject jsonObj = new JSONObject();
+				jsonObj = (JSONObject) jsonArr.get(j);
+				String sGUID = (String) jsonObj.get("newFileName");
+				String fileName = (String) jsonObj.get("pFileName");
 				
-				File file = new File(pDirPath + commonUtil.separator + filePath + commonUtil.separator + sGUID + ";" + fileName);
+				File file = new File(pDirPath + commonUtil.separator + filePath + commonUtil.separator + sGUID + fileName);
 				
 				file.delete();
-			}			
+			}
 		}
 
         logger.debug("tempUploadFileDelete ended");
@@ -2593,4 +2635,112 @@ public class EzCircularController extends EgovFileMngUtil {
 		logger.debug("circularConfirmPagingList ended");
 		return resultXML.toString();
 	}
+	
+	/**
+	 * 2018-07-12 김보미 - 모두저장(압축파일 내려받기)
+	 */
+	@RequestMapping(value="/ezCommunity/downloadAttachAll.do", produces="text/plain; charset=UTF-8")
+	public void downloadAttachAll(@CookieValue("loginCookie") String loginCookie, Locale locale, 
+			HttpServletRequest request, HttpServletResponse response) throws Exception {
+		logger.debug("downloadAttachAll started.");
+		
+		LoginVO userInfo = commonUtil.userInfo(loginCookie);
+		
+		String filePath = request.getParameter("filePath");
+		String fileNames = request.getParameter("fileNames");
+		String fileNames2 = request.getParameter("fileNames2");
+		String realPath = commonUtil.getRealPath(request);
+		String uploadFilePath = commonUtil.getUploadPath("upload_circular.ROOT", userInfo.getTenantId());
+		String tempFileUploadPath = realPath + uploadFilePath + commonUtil.separator + "tempUploadFile";
+		String guid = UUID.randomUUID().toString();
+		String pDirTempPath = tempFileUploadPath + commonUtil.separator + guid;
+		int bufferSize = 4096;
+
+		logger.debug("filePath : " + filePath + " | fileName : " + fileNames);
+
+		String fullFilePath = realPath + uploadFilePath + commonUtil.separator + "uploadFile" + filePath;
+
+		logger.debug("fullFilePath : " + fullFilePath);
+		
+		ZipOutputStream zos = null;
+		String downFileName = "";
+		
+		try {
+			File tempFile = new File(pDirTempPath + commonUtil.separator + ".zip");
+			
+			if (tempFile.exists()) {
+				tempFile.delete();
+			}
+			
+			tempFile = new File(tempFileUploadPath);
+			
+			if (!tempFile.exists()) {
+				tempFile.mkdirs();
+			}
+			
+			zos = new ZipOutputStream(new FileOutputStream(pDirTempPath + ".zip"), Charset.forName("utf-8"));
+				
+			JSONParser jp = new JSONParser();
+			JSONArray fileNamesArr = (JSONArray)jp.parse(fileNames);
+			JSONArray fileNamesArr2 = (JSONArray)jp.parse(fileNames2);
+
+			downFileName = fileNamesArr2.get(0).toString() + " 외 " + (fileNamesArr2.size() - 1) + "개.zip";//zip파일명
+			
+			if (fileNamesArr.size() != 0) {// 파일이 있으면
+				for (int i = 0; i < fileNamesArr.size(); i++) { //파일 길이만큼
+					BufferedInputStream bis = null;
+					
+					try {
+				       File sourceFile = new File(fullFilePath + fileNamesArr.get(i).toString());
+	                   
+				        bis = new BufferedInputStream(new FileInputStream(sourceFile));
+				        ZipEntry zentry = new ZipEntry(fileNamesArr2.get(i).toString());
+				        zos.putNextEntry(zentry);
+				        
+				        byte[] buffer = new byte[bufferSize];
+				        int cnt = 0;
+				        while ((cnt = bis.read(buffer, 0, bufferSize)) != -1) {
+				            zos.write(buffer, 0, cnt);
+				        }
+				        zos.closeEntry();
+					} catch (IOException e) {
+						e.printStackTrace();
+					} finally {
+						if (bis != null) {
+							try {
+								bis.close();
+							} catch (Exception e) {
+								e.printStackTrace();
+							}
+						}
+					}
+				}
+				zos.flush();
+				zos.close();
+				zos = null;
+	
+				File file = new File(pDirTempPath + ".zip");
+				
+				if (file.exists()) {
+					downFile(request, response, pDirTempPath + ".zip", downFileName);
+					file.delete();
+				}
+			}
+		} catch (Exception e) {
+			File file = new File(pDirTempPath + ".zip");
+			
+			if (file.exists()) {
+				file.delete();
+			}
+		} finally {
+			if (zos != null) {
+				try {
+					zos.close();
+				} catch (Exception e) {
+				}
+			}
+		}
+		logger.debug("downloadAttachAll ended.");
+	}
+		
 }
